WIP: TEMP-FIX! manually copy payload.js to dist
Problem: The `addDocketCounts` function specifies the name 'payload.js'
as the file that should be executed in the chrome tab when scraping
the page. However, when parcel builds the source files, the dist/ files
all have a unique hash added to them so the file becomes something
like 'payload.89274893.js'. I haven't figured out how to tell parcel
to update this filename though.

Solution: No idea. The problem is simiar to this issue, and
this might be an option, but it feels a little smelly to me:
https://github.com/parcel-bundler/parcel/issues/1668

For now: I'm copying the file over manually so that the file is
available as expected. Seems hacky, but I think it's currently working.