WIP: fix 'parcel watch' websocket error
When using 'parcel watch' to rebuild the extension as files change, the websocket that parcel generates when visiting the 'Filings' page was breaking.

The solution was to add the '--hmr-hostname' flag to the watch script.

This parcel issue explains the problem I was facing: https://github.com/parcel-bundler/parcel/issues/5254